Rename TileMapInfo into TileMapData (removing old version) and restore constructing from IntArray with masks for flags
